% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_markdown.R
\name{as_markdown}
\alias{as_markdown}
\title{Parse into Markdown}
\usage{
as_markdown(x, ...)
}
\arguments{
\item{x}{Object}

\item{...}{Further arguments passed to methods}
}
\description{
S3 Methods to parse different classes into Markdown.
}
