% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{fetchRdDB}
\alias{fetchRdDB}
\title{Make first letter capital}
\usage{
fetchRdDB(filebase, key = NULL)
}
\arguments{
\item{filebase}{The file path to the database (\code{.rdb} file), with no extension}

\item{key}{Keys to fetch}
}
\value{
character vector with capitalized first letters
}
\description{
Capitalize the first letter of every new word. Very simplistic approach.
}
