% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output_format.R
\name{md_document}
\alias{md_document}
\title{Output Format Markdown}
\usage{
md_document(
  file_ext = ".md",
  section_level = 1,
  date_format = "\%F",
  toc = FALSE
)
}
\arguments{
\item{file_ext}{File extension, if default output_file name}

\item{section_level}{Initial section level, subsection level \code{n} will
consequently be \code{section_level + n}}

\item{date_format}{See \code{format} argument of \code{\link[base:strptime]{base::strptime()}}}

\item{toc}{Currently not implemented. If table of contents should be included
in reference manual.}
}
\description{
Specifies markdown output format parameters.
}
