% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output_format.R
\name{output_format}
\alias{output_format}
\title{Output Format Definition}
\usage{
output_format(output_class_suffix, file_ext, parse_fun, ...)
}
\arguments{
\item{output_class_suffix}{Suffix to class name. E.g. \code{md} for
\code{output_format_md}.}

\item{file_ext}{File extension, if default output_file name.}

\item{parse_fun}{Parsing function, should be a generic S3 method.}

\item{...}{Named arguments inserted into list structure output and eventually
passed as arguments to \code{parse_fun}.}
}
\description{
Specifies output format parameters.
}
