% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parseRd.R
\name{parse_unknown_rd}
\alias{parse_unknown_rd}
\title{parse unknown rdfile}
\usage{
parse_unknown_rd(rdfile)
}
\arguments{
\item{rdfile}{either rd filename or entry in .rdb database (from \code{\link{fetchRdDB}})}
}
\value{
a named list with the parts of the Rd object that will be used for creating
a markdown file
}
\description{
\code{\link{ReferenceManual}} looks for actual .rd files, or in the case of already
compiled package, will read from .rdb database.  This function checks whether rdfile is
a filename or a string from the .rdb database, and parses accordingly.
}
\keyword{internal}
