\docType{package}
\name{Rd2markdown-package}
\alias{Rd2markdown-package}
\title{Convert Rd help files to markdown.}
\description{
  Functions to parse and convert Rd files to markdown.
}
\author{
  Jason Bryer \email{jason@bryer.org}
}
\seealso{
  \code{knitr}
}
\keyword{Rd}
\keyword{help}
\keyword{markdown}

