% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rd2markdown.R
\name{Rd2markdown}
\alias{Rd2markdown}
\title{Convert Rd help files to markdown}
\usage{
Rd2markdown(pkg, outfile, verbose = FALSE)
}
\arguments{
\item{pkg}{the package to generate files for.}

\item{outfile}{Output file path}

\item{verbose}{if \code{TRUE}, print messages as parsing.}
}
\description{
Converts the help topics in the given package to markdown.
}
\references{
Murdoch, D. (2010). 
          \href{http://developer.r-project.org/parseRd.pdf}{Parsing Rd files}.
}
\seealso{
Rd2markdown.rd
}

