% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rd2markdown.rd.R
\name{Rd2markdown.rd}
\alias{Rd2markdown.rd}
\title{Converts an Rd help file to markdown.}
\usage{
Rd2markdown.rd(rd, outfile, knitrOptions = c(message = FALSE, warning = FALSE,
  error = FALSE))
}
\arguments{
\item{rd}{an Rd data object.}

\item{outfile}{Filepath to output file (markdown file)}

\item{knitrOptions}{options to pass to \code{knitr} for running the examples.}
}
\value{
a character vector of length one where the element name is the topic
       name and the value is the filename.
}
\description{
This function will convert an Rd file to markdown.
}

