% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rd2md-package.R
\docType{package}
\name{Rd2md-package}
\alias{Rd2md-package}
\title{Convert Rd help files to markdown.}
\description{
Functions to parse and convert Rd files to markdown.
}
\author{
Julian Busch, copied from Jason Bryer \email{jason@bryer.org}
}
\seealso{
\code{knitr}
}
\keyword{Rd}
\keyword{help}
\keyword{markdown}

