% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rd2md}
\alias{rd2md}
\title{Translate an Rd string to its HTML output}
\usage{
rd2md(x, fragment = TRUE, ...)
}
\arguments{
\item{x}{Rd string. Backslashes must be double-escaped ("\\\\").}

\item{fragment}{logical indicating whether this represents a complete Rd file}

\item{...}{additional arguments for as_markdown}
}
\description{
Translate an Rd string to its HTML output
}
\examples{
rd2md("a\n\%b\nc")

rd2md("a & b")

rd2md("\\\\strong{\\\\emph{x}}")

}
