% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdfile-operations.R
\name{read_rdfile}
\alias{read_rdfile}
\title{Read .Rd file}
\usage{
read_rdfile(path, pkg_path = NULL, subclass = NULL)
}
\arguments{
\item{path}{File path of \verb{*.Rd}-file}

\item{pkg_path}{Package path, required to load macros, see
\code{\link[tools:loadRdMacros]{tools::loadPkgRdMacros()}}.}

\item{subclass}{Potential subclass of returned \code{rdfile} object}
}
\description{
And return object of class \code{rdfile}
}
