% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_ReferenceManual.R
\name{ReferenceManual}
\alias{ReferenceManual}
\title{Create Reference Manual Markdown}
\usage{
ReferenceManual(pkg_, outdir_ = getwd(), verbose_ = FALSE)
}
\arguments{
\item{pkg_}{Package name or path to directory}

\item{outdir_}{output dir where the reference manual markdown shall be written to}

\item{verbose_}{IF \code{TRUE} all messages and steps will be printed.}
}
\description{
This is a wrapper for a slightly
amended version of package Rd2markdown by jbryer (Github).
It takes the available functions and puts all into one document - the reference manual.
}

