% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{simpleCap}
\alias{simpleCap}
\title{Make first letter capital}
\usage{
simpleCap(x)
}
\arguments{
\item{x}{Character string}
}
\value{
character vector with capitalized first letters
}
\description{
Capitalize the first letter of every new word. Very simplistic approach.
}

