\name{stripWhite}
\alias{stripWhite}
\title{Strip white space.}
\usage{
  stripWhite(x)
}
\arguments{
  \item{x}{character to strip white space from.}
}
\value{
  a character with white space stripped.
}
\description{
  Strip white space (spaces only) from the beginning and
  end of a character.
}

