% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parseTag.R
\name{parseTag}
\alias{parseTag}
\title{Tag from Rd Format to Markdown}
\usage{
parseTag(
  x,
  pre = character(),
  post = character(),
  stripNewline = TRUE,
  stripWhite = TRUE,
  stripTab = TRUE
)
}
\arguments{
\item{x}{element from an \code{Rd} class.}

\item{pre}{String to prepend to the parsed tag.}

\item{post}{String to append to the parsed tag.}

\item{stripNewline}{Logical indicating whether to strip new line characters.}

\item{stripWhite}{Logical indicating whether to strip white space.}

\item{stripTab}{Logical indicating whether to strip tab characters.}
}
\description{
This function will convert an Rd element to markdown format. 
Note that links are only supported within a markdown document, referenced with #.
Use \code{href} to link to external sources.
}
