% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rd2markdown.R
\name{Rd2markdown}
\alias{Rd2markdown}
\title{Rd file to markdown}
\usage{
Rd2markdown(rdfile, outfile, append = FALSE, section = "#",
  subsection = "##", run.examples = FALSE, Rmd = F)
}
\arguments{
\item{rdfile}{Filepath to an .Rd file or an \code{Rd} object to parse.}

\item{outfile}{Filepath to output file (markdown file).}

\item{append}{If outfile exists, append to existing content.}

\item{section}{header tag.}

\item{subsection}{header tag.}

\item{run.examples}{logical. should examples be run?}
}
\value{
Parsed Rd as named list
}
\description{
This function converts an Rd file into markdown format.
}
\examples{
## give a markdown source file
rdfile = "~/git/MyPackage/man/myfun.Rd"
## specify, where markdown shall be stored
outfile = "/var/www/html/R_Web_app/md/myfun.md"
## create markdown
## Rd2markdown(rdfile = rdfile, outfile = outfile)
}
