% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rd2markdown.R
\name{Rd2markdown}
\alias{Rd2markdown}
\title{Rd file to markdown}
\usage{
Rd2markdown(rd, outfile, front.matter = "")
}
\arguments{
\item{rd}{an Rd data object.}

\item{outfile}{Filepath to output file (markdown file)}

\item{front.matter}{String}
}
\value{
a character vector of length one where the element name is the topic
name and the value is the filename.
}
\description{
This function converts an Rd file into markdown format.
}

