% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rd2md-package.R
\docType{package}
\name{Rd2md-package}
\alias{Rd2md-package}
\title{Reference Manual in Markdown}
\description{
Functions to parse and convert Rd files to markdown format reference manual.
}
\details{
Forked from Julian Busch <jb@quants.ch> Rd2md package.
}
\seealso{
\href{https://CRAN.R-project.org/package=Rd2md}{Rd2md in CRAN}
}
\author{
Poorani Subramanian
}
\keyword{Rd}
\keyword{help}
\keyword{manual}
\keyword{markdown}
\keyword{reference}
