% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rd2md-package.R
\docType{package}
\name{Rd2md-package}
\alias{Rd2md-package}
\title{Reference Manual in Markdown}
\description{
Function to parse and convert Rd files to markdown format reference manual.
}
\author{
Julian Busch <jb@quants.ch>
}
\seealso{
\code{knitr}
}
\keyword{Rd}
\keyword{help}
\keyword{manual}
\keyword{markdown}
\keyword{reference}

