% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_ReferenceManual.R
\name{ReferenceManual}
\alias{ReferenceManual}
\title{Create Reference Manual Markdown}
\usage{
ReferenceManual(pkg = getwd(), outdir = getwd(), verbose = FALSE)
}
\arguments{
\item{pkg}{Full path to package directory. Default value is the working directory}

\item{outdir}{Output directory where the reference manual markdown shall be written to}

\item{verbose}{If \code{TRUE} all messages and process steps will be printed.}
}
\description{
This is a wrapper for a slightly amended version of package Rd2markdown by 
[jbryer]('https://github.com/jbryer/Rd2markdown').
It takes slightly amended versions of the available functions, so that the manuals are
taken from source instead of from the libraries. 
The result is the reference manual in markdown format.
}

